Schedule a thread to collect cpu usage

+++ dubbo-monitor/dubbo-monitor-default/src/main/java/org/apache/dubbo/monitor/dubbo/CpuUsageListener.java
@@ -1,7 +1,5 @@
 package org.apache.dubbo.monitor.dubbo;
 
-import java.util.concurrent.atomic.AtomicLong;
-
 public interface CpuUsageListener {
-    void cpuChanged(AtomicLong cpu);
+    void cpuChanged(Float cpu);
 }
